plugins: Add Linux PSI plugin
PSI is the most modern and comprehensive way to measure resource
contention on Linux servers, see
https://www.kernel.org/doc/html/latest/accounting/psi.html. It's a
relatively mature kernel interface and has already proved itself to be
very useful finding and debugging issues at scale, and it makes sense to
have in the munin main repository.

Metrics present as percent of time spent doing "wasted" work due to
contention, like having to refault pages from disk. This avoids having
to guess when a workload is approaching the limit based on things like
the used memory and memory composition, which can be highly misleading.